requirements.txt
boto3
python-dotenv
pillow
tqdm
